(#26) Merge Parser and Lexer
- The lexer is now part of the parser, reducing the compilation time for
  big files when syntax errors are in the beginning of the program
- Close #26
